+++ src/main/java/org/swisspush/redisques/util/RedisquesAPI.java
@@ -54,10 +54,10 @@ public class RedisquesAPI {
     public static final String PER_QUEUE_CONFIG_ENQUEUE_DELAY_FACTOR_MILLIS =  "enqueueDelayFactorMillis";
     public static final String PER_QUEUE_CONFIG_ENQUEUE_MAX_DELAY_MILLIS =  "enqueueMaxDelayMillis";
     public static final String PER_QUEUE_CONFIG_MAX_QUEUE_ENTRIES =  "maxQueueEntries";
+    public static final String PER_QUEUE_CONFIG_MAX_QUEUE_PATROL_LIMIT =  "queuePatrolLimit";
     public static final String PER_QUEUE_CONFIG_MAXIMUM_ITEM_IN_BATCH_DISPATCH =  "maximumItemInBatchDispatch";
     public static final String PER_QUEUE_CONFIG_MINIMUM_ITEM_IN_BATCH_DISPATCH =  "minimumItemInBatchDispatch";
     public static final String PER_QUEUE_CONFIG_MAX_BATCH_DISPATCH_WAIT_TIMEOUT =  "maxBatchItemDispatchWaitTimeout";
-
     public static final String PER_QUEUE_CONFIG_NAME =  "configName";
     public static final String PER_QUEUE_CONFIG_PATTERN =  "pattern";
 
@@ -94,7 +94,8 @@ public class RedisquesAPI {
         monitor(null),
         setPerQueueConfiguration(null),
         getPerQueueConfiguration(null),
-        deleteQueueConfiguration(null);
+        deleteQueueConfiguration(null),
+        getQueuesSizeStatistics(null);
 
 
         private final String legacyName;
@@ -304,6 +305,14 @@ public class RedisquesAPI {
             return buildOperation(QueueOperation.getQueuesStatistics, new JsonObject().put(FILTER, filterPattern));
         }
         return buildGetQueuesStatisticsOperation();
+    }
+
+    /**
+     * Retrieve all queues size from queue statistics
+     *
+     */
+    public static JsonObject buildGetQueuesSizeStatisticsOperation() {
+            return buildOperation(QueueOperation.getQueuesSizeStatistics, new JsonObject());
     }
 
     /**