+++ src/test/java/org/swisspush/redisques/util/RedisquesConfigurationTest.java
@@ -65,7 +65,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(config.getQueueStatsRequestQuotaAcquireRetryTimeMs(), -1);
         testContext.assertEquals(config.getActiveQueueRegRefreshReqQuotaAcquireRetryTimeMs(), -1);
         testContext.assertFalse(config.getTcpKeepAlive());
-        testContext.assertEquals(config.getGlobalQueuePatrol(), -1);
         testContext.assertEquals(config.getRedisReplicasType(), RedisReplicas.NEVER);
     }
 
@@ -110,7 +109,6 @@ public class RedisquesConfigurationTest {
                 .checkQueueRequestsQuotaAcquireRetryTimeMs(4)
                 .queueStatsRequestQuotaAcquireRetryTimeMs(5)
                 .tcpKeepAlive(true)
-                .globalQueuePatrol(9090)
                 .redisReplicasType(RedisReplicas.ALWAYS)
                 .build();
 
@@ -156,7 +154,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(config.getQueueStatsRequestQuotaAcquireRetryTimeMs(), 5);
         testContext.assertEquals(config.getActiveQueueRegRefreshReqQuotaAcquireRetryTimeMs(), 3);
         testContext.assertTrue(config.getTcpKeepAlive());
-        testContext.assertEquals(config.getGlobalQueuePatrol(), 9090);
         testContext.assertEquals(config.getRedisReplicasType(), RedisReplicas.ALWAYS);
 
         // queue configurations
@@ -213,7 +210,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(json.getInteger(PROP_ACTIVE_QUEUE_REFRESH_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS), -1);
         testContext.assertEquals(json.getInteger(PROP_QUEUE_STATS_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS), -1);
         testContext.assertFalse(json.getBoolean(PROP_REDIS_CONNECTION_KEEP_ALIVE));
-        testContext.assertEquals(json.getInteger(PROP_GLOBAL_QUEUE_PATROL), -1);
         testContext.assertEquals(RedisReplicas.valueOf(json.getString(PROP_REDIS_REPLICAS_TYPE)), RedisReplicas.NEVER);
     }
 
@@ -257,7 +253,6 @@ public class RedisquesConfigurationTest {
                 .checkQueueRequestsQuotaAcquireRetryTimeMs(4)
                 .queueStatsRequestQuotaAcquireRetryTimeMs(5)
                 .tcpKeepAlive(true)
-                .globalQueuePatrol(8080)
                 .redisReplicasType(RedisReplicas.SHARE)
                 .build();
 
@@ -307,7 +302,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(json.getInteger(PROP_CHECK_QUEUE_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS), 4);
         testContext.assertEquals(json.getInteger(PROP_QUEUE_STATS_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS), 5);
         testContext.assertTrue(json.getBoolean(PROP_REDIS_CONNECTION_KEEP_ALIVE));
-        testContext.assertEquals(json.getInteger(PROP_GLOBAL_QUEUE_PATROL), 8080);
         testContext.assertEquals(RedisReplicas.valueOf(json.getString(PROP_REDIS_REPLICAS_TYPE)), RedisReplicas.SHARE);
 
         // queue configurations
@@ -366,7 +360,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(config.getRedisMonitoringReqQuotaAcquireRetryTimeMs(), -1);
         testContext.assertEquals(config.getCheckQueueRequestsQuotaAcquireRetryTimeMs(), -1);
         testContext.assertFalse(config.getTcpKeepAlive());
-        testContext.assertEquals(config.getGlobalQueuePatrol(), -1);
         testContext.assertEquals(config.getRedisReplicasType(), RedisReplicas.NEVER);
     }
 
@@ -420,7 +413,6 @@ public class RedisquesConfigurationTest {
         json.put(PROP_ACTIVE_QUEUE_REFRESH_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS, 3);
         json.put(PROP_QUEUE_STATS_REQUEST_QUOTA_ACQUIRE_RETRY_TIME_MS, 4);
         json.put(PROP_REDIS_CONNECTION_KEEP_ALIVE, true);
-        json.put(PROP_GLOBAL_QUEUE_PATROL, 999);
         json.put(PROP_REDIS_REPLICAS_TYPE, RedisReplicas.ALWAYS);
 
         RedisquesConfiguration config = fromJsonObject(json);
@@ -464,7 +456,6 @@ public class RedisquesConfigurationTest {
         testContext.assertEquals(config.getRedisMonitoringReqQuotaAcquireRetryTimeMs(), 2);
         testContext.assertEquals(config.getCheckQueueRequestsQuotaAcquireRetryTimeMs(), 1);
         testContext.assertTrue(config.getTcpKeepAlive());
-        testContext.assertEquals(config.getGlobalQueuePatrol(), 999);
         testContext.assertEquals(config.getRedisReplicasType(), RedisReplicas.ALWAYS);
 
         // queue configurations